fix(telemetry): restore the wasm and Windows builds
Two build failures, one of them mine.

**wasm (regression from the previous commit).** `FileDownloadSession::finalize_with`
put `crate::telemetry::Outcome` in its *signature*, but the whole telemetry
module was gated to non-wasm, so `xet-data` stopped compiling for
wasm32-unknown-unknown. Gating the method instead would have pushed `#[cfg]`
onto every caller that merely names an outcome, so the vocabulary moves to a
new always-compiled `telemetry::outcome` module: `Outcome`, `ERROR_CLASS_NONE`,
`error_class`, `classify_error`, `outcome_for_class`. `emit` and `payload` stay
gated, since those are what actually depend on `TransferTelemetry`.

This is a pure move; `Outcome` never appeared in the serialized payload
(`CommonMetrics::outcome` is a `&'static str` produced by `as_str()`), so the
generated schema is unchanged and the compatibility gate in the follow-up PR
still passes untouched.

`XetDownloadStreamGroup::finish_blocking` is also gated to non-wasm, matching
the other `_blocking` methods on that type - `bridge_sync` does not exist there.
The async `finish` remains available on every target.

**Windows (pre-existing, from the simulation-route commit).** The `not(unix)`
`LocalTestServer` initializer was never given the `telemetry_docs` field, which
is ungated on the struct, so `build_and_test-win` failed with E0063 on every
push since. Both initializers now match their field sets exactly.

Verified locally against what CI actually runs: both wasm crates check for
wasm32-unknown-unknown, `cargo clippy -r -- -D warnings` is clean for the
workspace and for hf_xet, nightly rustfmt is clean, and the full test suite
passes. The Windows path could not be cross-compiled here (aws-lc-sys needs a
Windows C toolchain), so it was verified by comparing each initializer's field
set against the struct definition - which is exactly what E0063 checks.
