+++ xet_data/src/telemetry/mod.rs
@@ -7,16 +7,27 @@
 //!
 //! [`DeduplicationMetrics`]: crate::deduplication::DeduplicationMetrics
 //! [`GroupProgressReport`]: crate::progress_tracking::GroupProgressReport
+//!
+//! # Targets
+//!
+//! Only [`outcome`] compiles everywhere. Everything else depends on `TransferTelemetry`, which
+//! does not exist on wasm, so it is gated - matching `xet_client::cas_client::telemetry`. The
+//! outcome vocabulary stays ungated because it appears in `FileDownloadSession`'s public
+//! signatures, and gating it would push `#[cfg]` onto every caller that merely names an outcome.
 
+#[cfg(not(target_family = "wasm"))]
 mod emit;
+mod outcome;
+#[cfg(not(target_family = "wasm"))]
 mod payload;
 
+#[cfg(not(target_family = "wasm"))]
 pub(crate) use emit::{
     UploadSnapshot, emit_download_abandoned, emit_download_terminal, emit_upload_abandoned, emit_upload_terminal,
     start_download_heartbeat, start_upload_heartbeat,
 };
-pub use emit::{classify_error, outcome_for_class};
+pub use outcome::{ERROR_CLASS_NONE, Outcome, classify_error, error_class, outcome_for_class};
+#[cfg(not(target_family = "wasm"))]
 pub use payload::{
-    CommonInputs, CommonMetrics, DownloadMetrics, ERROR_CLASS_NONE, Outcome, TELEMETRY_SCHEMA_VERSION,
-    TransferIdentity, UploadMetrics, error_class,
+    CommonInputs, CommonMetrics, DownloadMetrics, TELEMETRY_SCHEMA_VERSION, TransferIdentity, UploadMetrics,
 };
